+++ backend/src/main/java/meetween/backend/place/dto/request/PlaceInfoRequest.java
@@ -1,23 +1,31 @@
-package meetween.backend.place.dto;
+package meetween.backend.place.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import meetween.backend.place.domain.Cafe;
 import meetween.backend.place.domain.Restaurant;
 
 import java.util.List;
 import java.util.Optional;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class RestaurantInfoRequest {
+public class PlaceInfoRequest {
 
     @JsonProperty("list_total_count")
     private int totalCount;
     @JsonProperty("row")
-    private List<RestaurantApiRequest> items;
+    private List<PlaceApiRequest> items;
 
     public List<Restaurant> toRestaurants() {
         return items.stream()
-                .map(RestaurantApiRequest::toRestaurant)
+                .map(PlaceApiRequest::toRestaurant)
+                .flatMap(Optional::stream)
+                .toList();
+    }
+
+    public List<Cafe> toCafes() {
+        return items.stream()
+                .map(PlaceApiRequest::toCafe)
                 .flatMap(Optional::stream)
                 .toList();
     }
